Add separate class for person
