+++ app.rb
@@ -1,5 +1,4 @@
-require_relative './student'
-require_relative './teacher'
+require_relative './people_options.rb'
 require_relative './book'
 require_relative './rental'
 
@@ -7,8 +6,30 @@ class App
   def initialize(menu)
     @menu = menu
     @books = []
-    @people = []
+    @people_options = PeopleOptions.new
     @rentals = []
+  end
+
+  def select_option(user_input)
+    case user_input
+    when 1
+      list_all_books
+    when 2
+      @people_options.list_all_people
+    when 3
+      @people_options.create_person
+    when 4
+      create_book
+    when 5
+      create_rental
+    when 6
+      list_all_rentals
+    when 7
+      exit
+    else
+      puts 'Please select the number from the given list!'
+    end
+    @menu.show_menu
   end
 
   def list_all_books
@@ -17,56 +38,6 @@ class App
     else
       @books.each { |book| puts "Title: #{book.title}, Author: #{book.author}" }
     end
-    @menu.show_menu
-  end
-
-  def list_all_people
-    if @people.empty?
-      puts 'No added person!'
-    else
-      @people.each { |person| puts "[#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}" }
-    end
-    @menu.show_menu
-  end
-
-  def create_student
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = gets.chomp.downcase == 'y'
-    print 'Classroom: '
-    classroom = gets.chomp
-    @people.push(Student.new(classroom, age, name: name, parent_permission: parent_permission))
-    puts
-    puts 'Person created successfuly'
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Specialization: '
-    specialization = gets.chomp
-    @people.push(Teacher.new(specialization, age, name: name))
-    puts
-    puts 'Person created successfuly'
-  end
-
-  def create_person
-    puts 'Do you want to create a student(1) or a teacher(2)? [Input the number]'
-    person = gets.chomp
-    case person
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    else
-      puts 'Please add either 1 or 2!'
-    end
-
     @menu.show_menu
   end
